Refactor ParsingError subclasses:
- Relocated ParsingError subclasses (e.g., `UnsupportedJointType`, `MalformedPose`, etc.) from exceptions to adapters/gazebo.
- Updated DAO mappings in ORM to reference new locations.
- Removed redundant methods for raising the robot torso in transport actions.
- Adjusted related unit tests to match the new error hierarchy and structure.
