+++ semantic_digital_twin/src/semantic_digital_twin/exceptions.py
@@ -2,7 +2,7 @@ from __future__ import annotations, absolute_import
 
 from dataclasses import dataclass, field, Field
 from pathlib import Path
-from typing import Dict, Set, Any
+from typing import Dict, Set
 from uuid import UUID
 
 import mujoco
@@ -1042,152 +1042,6 @@ class PathResolutionError(ParsingError):
 
     def suggest_correction(self) -> str:
         return ""
-
-
-@dataclass
-class MissingRootElement(ParsingError):
-    """
-    Raised when a description file contains none of the expected root elements.
-    """
-
-    expected_elements: List[str] = field(kw_only=True, default_factory=list)
-    """
-    The element names that were searched for.
-    """
-
-    def error_message(self) -> str:
-        return f"None of the expected root elements {', '.join(self.expected_elements)} were found."
-
-    def suggest_correction(self) -> str:
-        return ""
-
-
-@dataclass
-class MalformedPose(ParsingError):
-    """
-    Raised when a pose does not consist of a position and roll-pitch-yaw triple.
-    """
-
-    text: str = field(kw_only=True)
-    """
-    The pose text that could not be interpreted.
-    """
-
-    def error_message(self) -> str:
-        return f"Pose '{self.text}' does not contain 6 values."
-
-    def suggest_correction(self) -> str:
-        return "Write the pose as 'x y z roll pitch yaw'."
-
-
-@dataclass
-class UnsupportedJointType(ParsingError):
-    """
-    Raised when a parsed joint uses a type that has no connection counterpart.
-    """
-
-    joint_name: str = field(kw_only=True)
-    """
-    The name of the joint that could not be mapped.
-    """
-
-    joint_type: str = field(kw_only=True)
-    """
-    The joint type that is not supported.
-    """
-
-    supported_types: List[str] = field(kw_only=True, default_factory=list)
-    """
-    The joint types that can be mapped to connections.
-    """
-
-    def error_message(self) -> str:
-        return f"Joint '{self.joint_name}' has unsupported type '{self.joint_type}'."
-
-    def suggest_correction(self) -> str:
-        if not self.supported_types:
-            return ""
-        return f"Use one of the supported types: {', '.join(sorted(self.supported_types))}."
-
-
-@dataclass
-class UnsupportedGeometryType(ParsingError):
-    """
-    Raised when a parsed geometry uses a shape that has no counterpart.
-    """
-
-    geometry_type: str = field(kw_only=True)
-    """
-    The geometry type that is not supported.
-    """
-
-    supported_types: List[str] = field(kw_only=True, default_factory=list)
-    """
-    The geometry types that can be mapped to shapes.
-    """
-
-    def error_message(self) -> str:
-        return f"Unsupported geometry type '{self.geometry_type}'."
-
-    def suggest_correction(self) -> str:
-        if not self.supported_types:
-            return ""
-        return f"Use one of the supported types: {', '.join(sorted(self.supported_types))}."
-
-
-@dataclass
-class UnsupportedPoseReference(ParsingError):
-    """
-    Raised when a pose is expressed relative to a named frame.
-
-    Poses are only supported with their default reference, which is the frame of the
-    element the pose belongs to.
-    """
-
-    attribute: str = field(kw_only=True)
-    """
-    The attribute carrying the frame reference, ``frame`` or ``relative_to``.
-    """
-
-    reference: str = field(kw_only=True)
-    """
-    The referenced frame.
-    """
-
-    def error_message(self) -> str:
-        return (
-            f"Pose is expressed relative to frame '{self.reference}' via "
-            f"'{self.attribute}', but only default frame semantics are supported."
-        )
-
-    def suggest_correction(self) -> str:
-        return f"Express the pose in its default frame and drop the '{self.attribute}' attribute."
-
-
-@dataclass
-class UnsupportedAxisReference(ParsingError):
-    """
-    Raised when a joint axis is expressed in a frame other than the joint frame.
-    """
-
-    joint_name: str = field(kw_only=True)
-    """
-    The name of the joint whose axis could not be interpreted.
-    """
-
-    reference: str = field(kw_only=True)
-    """
-    The frame the axis is expressed in.
-    """
-
-    def error_message(self) -> str:
-        return (
-            f"Axis of joint '{self.joint_name}' is expressed in '{self.reference}', "
-            f"but only the joint frame is supported."
-        )
-
-    def suggest_correction(self) -> str:
-        return "Express the axis in the joint frame."
 
 
 @dataclass
